requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorlog==6.9.0
cron-descriptor==1.4.5
distlib==0.4.0
Django==5.2.5
django-appconf==1.1.0
django-browser-reload==1.18.0
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-imagekit==5.0.0
django-redis==6.0.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-standardized-errors==0.15.0
drf-yasg==1.21.10
filelock==3.18.0
flake8==7.3.0
identify==2.6.13
idna==3.10
inflection==0.5.1
isort==6.0.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pilkit==3.0
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.33.1
wcwidth==0.2.13
whitenoise==6.9.0
